+++ CHANGELOG.md
@@ -9,6 +9,7 @@ See [Upgrading] for details how to upgrade.
 - Use Doctrine PDO connection for legacy PdoAdapter, #901
 - Use `Symfony\Contracts\EventDispatcher\Event` instead of deprecated `Symfony\Component`, #902
 - Configure Symfony logger: Processors, MailHandler, #904
+- Use Doctrine from Symfony in Doctrine::getEntityManager, #905
 
 [3.9.3]: https://github.com/eventum/eventum/compare/v3.9.2...master